# backend/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/motherboard/all/": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get All With Pagination",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Motherboard"
                ],
                "summary": "Get All With Pagination",
                "parameters": [
                    {
                        "description": "Get All With Pagination",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PaginationInputWithFilter"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Motherboard response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/motherboard/create": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a Motherboard",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Motherboard"
                ],
                "summary": "Create a Motherboard",
                "parameters": [
                    {
                        "description": "Create a Motherboard",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateMotherboardRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Motherboard response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/motherboard/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Delete Motherboard",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Motherboard"
                ],
                "summary": "Delete Motherboard",
                "responses": {
                    "204": {
                        "description": "Motherboard response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/motherboard/get/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "GetMotherboardById",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Motherboard"
                ],
                "summary": "Get Motherboard By Id",
                "responses": {
                    "200": {
                        "description": "GetMotherboardById response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/motherboard/update/{id}": {
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update Motherboard",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Motherboard"
                ],
                "summary": "Update Motherboard",
                "parameters": [
                    {
                        "description": "Update Motherboard",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMotherboardRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Motherboard response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/optical-drive/all/": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get All With Pagination",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "OpticalDrive"
                ],
                "summary": "Get All With Pagination",
                "parameters": [
                    {
                        "description": "Get All With Pagination",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PaginationInputWithFilter"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OpticalDrive response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/optical-drive/create": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a OpticalDrive",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "OpticalDrive"
                ],
                "summary": "Create a OpticalDrive",
                "parameters": [
                    {
                        "description": "Create a OpticalDrive",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOpticalDriveRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OpticalDrive response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/optical-drive/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Delete OpticalDrive",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "OpticalDrive"
                ],
                "summary": "Delete OpticalDrive",
                "responses": {
                    "204": {
                        "description": "OpticalDrive response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/optical-drive/get/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "GetOpticalDriveById",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "OpticalDrive"
                ],
                "summary": "Get OpticalDrive By Id",
                "responses": {
                    "200": {
                        "description": "GetOpticalDriveById response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/optical-drive/update/{id}": {
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update OpticalDrive",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "OpticalDrive"
                ],
                "summary": "Update OpticalDrive",
                "parameters": [
                    {
                        "description": " Update OpticalDrive",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateOpticalDriveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OpticalDrive response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateMotherboardRequest": {
            "type": "object",
            "properties": {
                "formFactor": {
                    "type": "string"
                },
                "powerConsumption": {
                    "type": "number"
                }
            }
        },
        "dto.CreateOpticalDriveRequest": {
            "type": "object",
            "properties": {
                "manufacturer": {
                    "type": "string"
                },
                "powerConsumption": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "dto.Filter": {
            "type": "object",
            "properties": {
                "filterType": {
                    "type": "string"
                },
                "form": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "dto.PaginationInputWithFilter": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/dto.Filter"
                    }
                },
                "page_number": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "sort": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SortDto"
                    }
                }
            }
        },
...
        "dto.SortDto": {
            "type": "object",
            "properties": {
                "colId": {
                    "type": "string"
                },
                "string": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateMotherboardRequest": {
            "type": "object",
            "properties": {
                "formFactor": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateOpticalDriveRequest": {
            "type": "object",
            "properties": {
                "manufacturer": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "helper.Response": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                },
                "result": {},
                "statusCode": {
                    "$ref": "#/definitions/helper.ResultCode"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "helper.ResultCode": {
            "type": "integer",
            "enum": [
                0,
                40001,
                40002,
                40101,
                40301,
                40401,
                42901,
                42902,
                50001,
                50002
            ],
            "x-enum-varnames": [
                "Success",
                "ValidationError",
                "Error",
                "AuthError",
                "ForbiddenError",
                "NotFoundError",
                "LimiterError",
                "OtpLimiterError",
                "CustomRecovery",
                "InternalError"
            ]
        }
    }
}